+++ src/viur/core/tasks.py
@@ -1,65 +1,91 @@
+import abc
 import base64
+import datetime
+import functools
 import grpc
+import json
 import logging
 import os
 import pytz
 import requests
 import sys
+import time
 import traceback
-from datetime import datetime, timedelta
-from functools import wraps
 import typing as t
+from google import protobuf
 from google.cloud import tasks_v2
-from google.cloud.tasks_v2.services.cloud_tasks.transports import CloudTasksGrpcTransport
-from google.protobuf import timestamp_pb2
-from time import sleep
-
-import json
 from viur.core import current, db, errors, utils
 from viur.core.config import conf
 from viur.core.decorators import exposed, skey
 from viur.core.module import Module
 
 
-# class JsonKeyEncoder(json.JSONEncoder):
-def preprocessJsonObject(o):
-    """
-        Add support for Keys, Datetime, Bytes and db.Entities in deferred tasks.
-        This is not a subclass of json.JSONEncoder anymore, as db.Entites are a subclass of dict, which
-        is always handled from the json module itself.
-    """
-    if isinstance(o, db.Key):
-        return {".__key__": db.encodeKey(o)}
-    elif isinstance(o, datetime):
-        return {".__datetime__": o.astimezone(pytz.UTC).strftime("%d.%m.%Y %H:%M:%S")}
-    elif isinstance(o, bytes):
-        return {".__bytes__": base64.b64encode(o).decode("ASCII")}
-    elif isinstance(o, db.Entity):
-        return {".__entity__": preprocessJsonObject(dict(o)), ".__ekey__": db.encodeKey(o.key) if o.key else None}
-    elif isinstance(o, dict):
-        return {preprocessJsonObject(k): preprocessJsonObject(v) for k, v in o.items()}
-    elif isinstance(o, (list, tuple, set)):
-        return [preprocessJsonObject(x) for x in o]
-    else:
-        return o
+CUSTOM_OBJ = t.TypeVar("CUSTOM_OBJ")  # A JSON serializable object
 
 
-def jsonDecodeObjectHook(obj):
+class CustomEnvironmentHandler(abc.ABC):
+    @abc.abstractmethod
+    def serialize(self) -> CUSTOM_OBJ:
+        """Serialize custom environment data
+
+        This function must not require any parameters and must
+        return a JSON serializable object with the desired information.
+        """
+        ...
+
+    @abc.abstractmethod
+    def restore(self, obj: CUSTOM_OBJ) -> None:
+        """Restore custom environment data
+
+        This function will receive the object from :meth:`serialize` and should write
+        the information it contains to the environment of the deferred request.
+        """
+        ...
+
+
+def _preprocess_json_object(obj):
     """
-        Inverse to JsonKeyEncoder: Check if the object matches a custom ViUR type and recreate it accordingly
+        Add support for db.Key, datetime, bytes and db.Entity in deferred tasks,
+        and converts the provided obj into a special dict with JSON-serializable values.
+    """
+    if isinstance(obj, db.Key):
+        return {".__key__": db.encodeKey(obj)}
+    elif isinstance(obj, datetime.datetime):
+        return {".__datetime__": obj.astimezone(pytz.UTC).isoformat()}
+    elif isinstance(obj, bytes):
+        return {".__bytes__": base64.b64encode(obj).decode("ASCII")}
+    elif isinstance(obj, db.Entity):
+        # TODO: Support Skeleton instances as well?
+        return {
+            ".__entity__": _preprocess_json_object(dict(obj)),
+            ".__ekey__": db.encodeKey(obj.key) if obj.key else None
+        }
+    elif isinstance(obj, dict):
+        return {_preprocess_json_object(k): _preprocess_json_object(v) for k, v in obj.items()}
+    elif isinstance(obj, (list, tuple, set)):
+        return [_preprocess_json_object(x) for x in obj]
+
+    return obj
+
+
+def _decode_object_hook(obj):
+    """
+        Inverse for _preprocess_json_object, which is an object-hook for json.loads.
+        Check if the object matches a custom ViUR type and recreate it accordingly.
     """
     if len(obj) == 1:
-        if ".__key__" in obj:
-            return db.Key.from_legacy_urlsafe(obj[".__key__"])
-        elif ".__datetime__" in obj:
-            value = datetime.strptime(obj[".__datetime__"], "%d.%m.%Y %H:%M:%S")
-            return datetime(value.year, value.month, value.day, value.hour, value.minute, value.second, tzinfo=pytz.UTC)
-        elif ".__bytes__" in obj:
-            return base64.b64decode(obj[".__bytes__"])
+        if key := obj.get(".__key__"):
+            return db.Key.from_legacy_urlsafe(key)
+        elif date := obj.get(".__datetime__"):
+            return datetime.datetime.fromisoformat(date)
+        elif buf := obj.get(".__bytes__"):
+            return base64.b64decode(buf)
+
     elif len(obj) == 2 and ".__entity__" in obj and ".__ekey__" in obj:
-        r = db.Entity(db.Key.from_legacy_urlsafe(obj[".__ekey__"]) if obj[".__ekey__"] else None)
-        r.update(obj[".__entity__"])
-        return r
+        entity = db.Entity(db.Key.from_legacy_urlsafe(obj[".__ekey__"]) if obj[".__ekey__"] else None)
+        entity.update(obj[".__entity__"])
+        return entity
+
     return obj
 
 
@@ -91,7 +117,9 @@ if not conf.instance.is_dev_server or os.getenv("TASKS_EMULATOR") is None:
     taskClient = tasks_v2.CloudTasksClient()
 else:
     taskClient = tasks_v2.CloudTasksClient(
-        transport=CloudTasksGrpcTransport(channel=grpc.insecure_channel(os.getenv("TASKS_EMULATOR")))
+        transport=tasks_v2.services.cloud_tasks.transports.CloudTasksGrpcTransport(
+            channel=grpc.insecure_channel(os.getenv("TASKS_EMULATOR"))
+        )
     )
     queueRegion = "local"
 
@@ -147,7 +175,7 @@ class CallableTaskBase:
         """
             The actual code that should be run goes here.
         """
-        raise NotImplemented()
+        raise NotImplementedError()
 
 
 class TaskHandler(Module):
@@ -193,7 +221,7 @@ class TaskHandler(Module):
         """
         req = current.request.get().request
         self._validate_request()
-        data = json.loads(req.body, object_hook=jsonDecodeObjectHook)
+        data = json.loads(req.body, object_hook=_decode_object_hook)
         if data["classID"] not in MetaQueryIter._classCache:
             logging.error(f"""Could not continue queryIter - {data["classID"]} not known on this instance""")
         MetaQueryIter._classCache[data["classID"]]._qryStep(data)
@@ -214,13 +242,17 @@ class TaskHandler(Module):
                 f"""Task {req.headers.get("X-Appengine-Taskname", "")} is retried for the {retryCount}th time."""
             )
 
-        cmd, data = json.loads(req.body, object_hook=jsonDecodeObjectHook)
+        cmd, data = json.loads(req.body, object_hook=_decode_object_hook)
         funcPath, args, kwargs, env = data
         logging.debug(f"Call task {funcPath} with {cmd=} {args=} {kwargs=} {env=}")
 
         if env:
             if "user" in env and env["user"]:
                 current.session.get()["user"] = env["user"]
+
+                # Load current user into context variable if user module is there.
+                if user_mod := getattr(conf.main_app, "user", None):
+                    current.user.set(user_mod.getCurrentUser())
             if "lang" in env and env["lang"]:
                 current.language.set(env["lang"])
             if "transactionMarker" in env:
@@ -232,11 +264,7 @@ class TaskHandler(Module):
                     logging.info(f"""Executing task, transaction {env["transactionMarker"]} did succeed""")
             if "custom" in env and conf.tasks_custom_environment_handler:
                 # Check if we need to restore additional environmental data
-                assert isinstance(conf.tasks_custom_environment_handler, tuple) \
-                       and len(conf.tasks_custom_environment_handler) == 2 \
-                       and callable(conf.tasks_custom_environment_handler[1]), \
-                    "Your customEnvironmentHandler must be a tuple of two callable if set!"
-                conf.tasks_custom_environment_handler[1](env["custom"])
+                conf.tasks_custom_environment_handler.restore(env["custom"])
         if cmd == "rel":
             caller = conf.main_app
             pathlist = [x for x in funcPath.split("/") if x]
@@ -253,7 +281,7 @@ class TaskHandler(Module):
                 logging.exception(e)
                 raise errors.RequestTimeout()  # Task-API should retry
         elif cmd == "unb":
-            if not funcPath in _deferred_tasks:
+            if funcPath not in _deferred_tasks:
                 logging.error(f"Missed deferred task {funcPath=} ({args=},{kwargs=})")
             # We call the deferred function *directly* (without walking through the mkDeferred lambda), so ensure
             # that any hit to another deferred function will defer again
@@ -281,7 +309,7 @@ class TaskHandler(Module):
             periodicTaskName = task.periodicTaskName.lower()
             if interval:  # Ensure this task doesn't get called to often
                 lastCall = db.Get(db.Key("viur-task-interval", periodicTaskName))
-                if lastCall and utils.utcNow() - lastCall["date"] < timedelta(minutes=interval):
+                if lastCall and utils.utcNow() - lastCall["date"] < datetime.timedelta(minutes=interval):
                     logging.debug(f"Task {periodicTaskName!r} has already run recently - skipping.")
                     continue
             res = self.findBoundTask(task, conf.main_app)
@@ -339,7 +367,7 @@ class TaskHandler(Module):
         ):
             logging.critical("Detected an attempted XSRF attack. This request did not originate from Task Queue.")
             raise errors.Forbidden()
-        if require_cron and "X-Appengine-Cron" not in req.request.headers:
+        if require_cron and "X-Appengine-Cron" not in req.headers:
             logging.critical('Detected an attempted XSRF attack. The header "X-AppEngine-Cron" was not set.')
             raise errors.Forbidden()
         if require_taskname and "X-AppEngine-TaskName" not in req.headers:
@@ -384,7 +412,7 @@ TaskHandler.vi = True
 TaskHandler.html = True
 
 
-## Decorators ##
+# Decorators
 
 def retry_n_times(retries: int, email_recipients: None | str | list[str] = None,
                   tpl: None | str = None) -> t.Callable:
@@ -405,7 +433,7 @@ def retry_n_times(retries: int, email_recipients: None | str | list[str] = None,
         <pre>{{traceback|escape}}</pre>"""
 
     def outer_wrapper(func):
-        @wraps(func)
+        @functools.wraps(func)
         def inner_wrapper(*args, **kwargs):
             try:
                 retry_count = int(current.request.get().request.headers.get("X-Appengine-Taskretrycount", -1))
@@ -456,7 +484,7 @@ def retry_n_times(retries: int, email_recipients: None | str | list[str] = None,
 
 def noRetry(f):
     """Prevents a deferred Function from being called a second time"""
-    logging.warning(f"Use of `@noRetry` is deprecated; Use `@retry_n_times(0)` instead!", stacklevel=2)
+    logging.warning("Use of `@noRetry` is deprecated; Use `@retry_n_times(0)` instead!", stacklevel=2)
     return retry_n_times(0)(f)
 
 
@@ -513,7 +541,7 @@ def CallDeferred(func: t.Callable) -> t.Callable:
             # Run tasks inline
             logging.debug(f"{func=} will be executed inline")
 
-            @wraps(func)
+            @functools.wraps(func)
             def task():
                 if self is __undefinedFlag_:
                     return func(*args, **kwargs)
@@ -579,16 +607,12 @@ def CallDeferred(func: t.Callable) -> t.Callable:
 
             if conf.tasks_custom_environment_handler:
                 # Check if this project relies on additional environmental variables and serialize them too
-                assert isinstance(conf.tasks_custom_environment_handler, tuple) \
-                       and len(conf.tasks_custom_environment_handler) == 2 \
-                       and callable(conf.tasks_custom_environment_handler[0]), \
-                    "Your customEnvironmentHandler must be a tuple of two callable if set!"
-                env["custom"] = conf.tasks_custom_environment_handler[0]()
+                env["custom"] = conf.tasks_custom_environment_handler.serialize()
 
             # Create task description
             task = tasks_v2.Task(
                 app_engine_http_request=tasks_v2.AppEngineHttpRequest(
-                    body=json.dumps(preprocessJsonObject((command, (funcPath, args, kwargs, env)))).encode("UTF-8"),
+                    body=json.dumps(_preprocess_json_object((command, (funcPath, args, kwargs, env)))).encode("UTF-8"),
                     http_method=tasks_v2.HttpMethod.POST,
                     relative_uri=taskargs["url"],
                     app_engine_routing=tasks_v2.AppEngineRouting(
@@ -602,10 +626,10 @@ def CallDeferred(func: t.Callable) -> t.Callable:
             # Set a schedule time in case eta (absolut) or countdown (relative) was set.
             eta = taskargs.get("eta")
             if seconds := taskargs.get("countdown"):
-                eta = utils.utcNow() + timedelta(seconds=seconds)
+                eta = utils.utcNow() + datetime.timedelta(seconds=seconds)
             if eta:
                 # We must send a Timestamp Protobuf instead of a date-string
-                timestamp = timestamp_pb2.Timestamp()
+                timestamp = protobuf.timestamp_pb2.Timestamp()
                 timestamp.FromDatetime(eta)
                 task.schedule_time = timestamp
 
@@ -617,9 +641,9 @@ def CallDeferred(func: t.Callable) -> t.Callable:
             logging.info(f"Created task {func.__name__}.{func.__module__} with {args=} {kwargs=} {env=}")
 
     global _deferred_tasks
-    _deferred_tasks[f"{func.__name__,}.{func.__module__}"] = func
+    _deferred_tasks[f"{func.__name__}.{func.__module__}"] = func
 
-    @wraps(func)
+    @functools.wraps(func)
     def wrapper(*args, **kwargs):
         return make_deferred(func, *args, **kwargs)
 
@@ -632,7 +656,7 @@ def callDeferred(func):
     """
     import logging, warnings
 
-    msg = f"Use of @callDeferred is deprecated, use @CallDeferred instead."
+    msg = "Use of @callDeferred is deprecated, use @CallDeferred instead."
     logging.warning(msg, stacklevel=3)
     warnings.warn(msg, stacklevel=3)
 
@@ -654,7 +678,7 @@ def PeriodicTask(interval: int = 0, cronName: str = "default") -> t.Callable:
         if fn.__name__.startswith("_"):
             raise RuntimeError("Periodic called methods cannot start with an underscore! "
                                f"Please rename {fn.__name__!r}")
-        if not cronName in _periodicTasks:
+        if cronName not in _periodicTasks:
             _periodicTasks[cronName] = {}
         _periodicTasks[cronName][fn] = interval
         fn.periodicTaskName = f"{fn.__module__}_{fn.__qualname__}".replace(".", "_").lower()
@@ -763,7 +787,7 @@ class QueryIter(object, metaclass=MetaQueryIter):
             parent=taskClient.queue_path(conf.instance.project_id, queueRegion, cls.queueName),
             task=tasks_v2.Task(
                 app_engine_http_request=tasks_v2.AppEngineHttpRequest(
-                    body=json.dumps(preprocessJsonObject(qryDict)).encode("UTF-8"),
+                    body=json.dumps(_preprocess_json_object(qryDict)).encode("UTF-8"),
                     http_method=tasks_v2.HttpMethod.POST,
                     relative_uri="/_tasks/queryIter",
                     app_engine_routing=tasks_v2.AppEngineRouting(
@@ -795,7 +819,7 @@ class QueryIter(object, metaclass=MetaQueryIter):
             try:
                 cls.handleEntry(item, qryDict["customData"])
             except:  # First exception - we'll try another time (probably/hopefully transaction collision)
-                sleep(5)
+                time.sleep(5)
                 try:
                     cls.handleEntry(item, qryDict["customData"])
                 except Exception as e:  # Second exception - call error_handler